\name{asht-package}
\alias{asht-package}
\alias{asht}
\docType{package}
\title{
Applied Statistical Hypothesis Tests
}
\description{
Test and confidence intervals for some applied statistical hypothesis tests.
}
\details{
\tabular{ll}{
Package: \tab asht\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2015-07-09\cr
License: \tab GPL-3\cr
}
A collection of statistical hypothesis tests, with a focus on non-asymptotic tests. Some tests are \code{\link{medianTest}} for exact tests and
confidence intervals about a median, \code{\link{quantileTest}} which generalizes \code{\link{medianTest}} for other quantiles besides the median, and \code{\link{signTest}} to run the exact sign test.
}
\author{
Michael P. Fay

Maintainer: Michael P. Fay <mfay@niaid.nih.gov>
}
%\references{}
\keyword{ package }
\keyword{htest}
%\seealso{~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~}
%\examples{~~ simple examples of the most important functions ~~}
